README first pass
This is to have a central repository for my work with the Kaggle dataset titled "OKCupid Profiles". 

When I picked up this dataset, I was struck by all of the fascinating tags that are attached to the samples (age, height, body type, whether the individual drinks/smokes/does drugs, etc.), as well as the rather large volume of unstructured language data. From that, one question that I dearly wanted to answer is this: can an NLP-stack infer any of these qualities about an unknown author with no other input other than the author's raw text?

I hope to lay out my methodology and findings in an accessible, reproducible format.